gcc compile
git-svn-id: file:///.data/svn_convert/svn@6855 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ branches/branch-QString/plugins/icq/icqlists.cpp
@@ -332,7 +332,7 @@ void ICQClient::parseRosterItem(unsigned short type,
                     buddyHash.resize(size);
                     memcpy(buddyHash.data(), &tlv_buddyHash->Data()[2], size);
 
-                    if(memcmp(ba.data(), buddyHash.data(), min(ba.size(), size))) {
+                    if(memcmp(ba.data(), buddyHash.data(), min(ba.size(), (unsigned int)size))) {
                         data.owner.buddyHash.asBinary() = buddyHash;
                         data.owner.buddyID.asULong() = flag;
                         requestBuddy(&data.owner);
